requirements.txt
click==6.7
Django==2.1.1
djangorestframework==3.8.2
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
py-spy==0.1.4
pytz==2018.5
Werkzeug==0.14.1
